backend/requirements.txt
firebase-admin
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
python-dotenv
numpy
torch
sentence_transformers
requests
pydantic
aiohttp